Moved the Github OAuth authentication into an extension method in a subfolder to allow startup.cs to be more readable.
